requirements.txt
absl-py==0.9.0
asgiref==3.2.5
astor==0.8.1
backports-abc==0.5
cachetools==4.0.0
certifi==2019.11.28
Chameleon==3.6.2
chardet==3.0.4
Django==3.0
dulwich==0.19.14
gast==0.2.2
Genshi==0.7.3
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
html5lib==1.0.1
idna==2.9
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Mako==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
mpmath==1.1.0
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.2.0
Pillow==7.0.0
protobuf==3.11.3
psutil==5.6.7
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyperf==1.7.0
pyperformance==1.0.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Shapely==1.7.0
singledispatch==3.4.0.3
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.12
sqlparse==0.3.1
sympy==1.5
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tornado==6.0.3
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.12.1
